requirements.txt
abipy
aiida-abinit~=0.4.0
aiida-bigdft>=0.2.6
aiida-castep>=1.2.0a5
aiida-core[atomic_tools]~=2.1
aiida-cp2k~=1.3
aiida-fleur>=1.3.0
aiida-gaussian
aiida-nwchem>=2.1.0
aiida-orca
aiida-pseudo>=0.6.0
aiida-quantumespresso~=3.4,>=3.4.1
aiida-siesta>=1.2.0
aiida-vasp
pymatgen>=2022.1.20
numpy<1.24.0
sqlalchemy<1.4
ase!=3.20.*
pint~=0.16
masci-tools~=0.9
